Chore: Done UI for text overlay
