+++ app/src/main/java/com/densitech/scrollsmooth/ui/text/view/TextConfigEnum.kt
@@ -1,0 +1,12 @@
+package com.densitech.scrollsmooth.ui.text.view
+
+enum class TextConfigEnum {
+    COLOR,
+    FONT
+}
+
+enum class TextAlignmentEnum {
+    LEFT,
+    CENTER,
+    RIGHT
+}
