+++ app/src/main/java/com/densitech/scrollsmooth/ui/video_transformation/view/VideoTransformationScreen.kt
@@ -8,14 +8,35 @@ import androidx.activity.compose.BackHandler
 import androidx.annotation.OptIn
 import androidx.compose.foundation.background
 import androidx.compose.foundation.border
-import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.KeyboardArrowDown
-import androidx.compose.material3.*
-import androidx.compose.runtime.*
+import androidx.compose.material3.BottomSheetScaffold
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
+import androidx.compose.material3.SheetValue
+import androidx.compose.material3.rememberBottomSheetScaffoldState
+import androidx.compose.material3.rememberStandardBottomSheetState
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.DisposableEffect
+import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.runtime.collectAsState
+import androidx.compose.runtime.derivedStateOf
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableFloatStateOf
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.runtime.saveable.rememberSaveable
+import androidx.compose.runtime.setValue
+import androidx.compose.runtime.snapshotFlow
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
@@ -37,7 +58,7 @@ import com.densitech.scrollsmooth.ui.audio.AudioSelectionViewModel
 import com.densitech.scrollsmooth.ui.bottom_sheet.SheetCollapsed
 import com.densitech.scrollsmooth.ui.bottom_sheet.SheetContent
 import com.densitech.scrollsmooth.ui.bottom_sheet.SheetExpanded
-import com.densitech.scrollsmooth.ui.text.TextOverlayPreview
+import com.densitech.scrollsmooth.ui.text.view.TextOverlayView
 import com.densitech.scrollsmooth.ui.utils.DEFAULT_FRACTION
 import com.densitech.scrollsmooth.ui.utils.format
 import com.densitech.scrollsmooth.ui.video_creation.model.DTOLocalVideo
@@ -292,6 +313,7 @@ fun VideoTransformationScreen(
             MainPreviewContent(
                 exoPlayer = exoPlayer,
                 currentFraction = progress.floatValue,
+                isShowingTextOverlay = isShowTextingOverlay,
                 onBackClick = {
                     navController.popBackStack()
                 },
@@ -332,6 +354,16 @@ fun VideoTransformationScreen(
                     )
             )
         }
+
+        if (isShowTextingOverlay) {
+            TextOverlayView(
+                onDoneClick = { textOverlayParams ->
+                    isShowTextingOverlay = false
+                }, modifier = Modifier
+                    .padding(top = 50.dp)
+                    .fillMaxSize()
+            )
+        }
     }
 
     // Bottom sheet audio
@@ -348,12 +380,6 @@ fun VideoTransformationScreen(
                 videoTransformationViewModel.playVideo()
             }
         )
-    }
-
-    if (isShowTextingOverlay) {
-        TextOverlayPreview(onDoneClick = {
-            isShowTextingOverlay = false
-        })
     }
 }
